031/separate server fuzz logic refactor (#54)
* Move setup routes to another method

* Move business logic out

* Attempt #1 at separating server and fuzzing logic

* Fixed dir bug on start

* debug before comitting refactor

* Fix flake8

* Review fixes
