+++ .gitignore
@@ -118,3 +118,6 @@ dmypy.json
 
 # Database
 *.db
+
+# JetBrains IDEs
+**.idea
