--- hypothesisfuzzer/exceptions/__init__.py
@@ -1,4 +1,0 @@
-from .invalid_config_exception import (  # noqa
-    ConfigMissingOptionException,
-    InvalidConfigException
-)
