+++ hypothesisfuzzer/fuzzing/__init__.py
@@ -1,0 +1,1 @@
+from .repo_fuzzer import RepoFuzzer  # noqa
